add support for integers, in particular return types (which need to be properly parsed during code gen)

+++ src/ast.h
@@ -57,13 +57,13 @@ class Statement : public AST {
 
 
 /**
- * Expression class for numeric literals.
+ * Expression class for numeric integer literals.
  */
 class NumericalExpr : public Expr {
-  double value;
+  long value;
 
   public:
-    NumericalExpr(double value) : value(value) {}
+    NumericalExpr(long value) : value(value) {}
     llvm::Value *codegen() override;
 };
